+++ R/utility.R
@@ -23,8 +23,13 @@ getOS <- function(){
 
 # get directory from a paht
 fileFromPath <- function(x){
-  all <- unlist(strsplit(x, "/"))
-  paste(all[length(all)], collapse="/")
+  res <- rep(NA, length(x))
+  for(i in 1:length(x)){
+	all <- unlist(strsplit(x[i], "/"))
+	res[i] <- paste(all[length(all)], collapse="/")
+  }
+  names(res) <- names(x)
+  return(res)
 }
 
 # from: https://ryouready.wordpress.com/2008/12/18/generate-random-string-name/
@@ -61,7 +66,7 @@ detailedBounds <- function(x,y, xmin=-180, xmax=180, ymin=-90, ymax=90){
 #' @param ymax (\code{numeric}) Maximum value of y (latitude).
 #' @param out (\code{character}) Output format, either \code{"sf"} or \code{"sp"}. The default \code{"sf"} returns simple feature geometries, \code{"sp"} returns \code{SpatialPolygons} from the \code{sp} package.
 #' 
-#' @return A \code{SpatialPolygons} class object.
+#' @return An \code{sfc}-, or \code{SpatialPolygons}-class object.
 #' @examples
 #' # requires rgdal
 #' edge <- mapedge()