+++ tests/offline/test_offline_sf_multilines.R
@@ -4,8 +4,7 @@ library(rgplates)
 model <- chronosphere::fetch("paleomap", "model", datadir=file.path(wd, "data/chronosphere"), verbose=FALSE)
 
 # example 
-admin <- chronosphere::fetch("paleomap", "adminlines", datadir=file.path(wd, "data/chronosphere"), verbose=FALSE)
-admin <-  st_as_sf(admin)
+admin <- st_read("/mnt/sky/Dropbox/WorkSpace/2021-04-28_rgplates/data/Paleomap/paleomap_adminlines_19o/admin.shx", quiet=TRUE)
 
 ################################################################################
 # 2. Explicit WGS 84 CRS