+++ tests/offline/test_offline_sp_polygons_complex.R
@@ -4,7 +4,8 @@ library(rgplates)
 model <- chronosphere::fetch("paleomap", "model", datadir=file.path(wd, "data/chronosphere"), verbose=FALSE)
 
 # example 
-myers <- chronosphere::fetch("Myers-hotspots", datadir=file.path(wd, "data/chronosphere"), verbose=FALSE)
+myers <- chronosphere::fetch("myers-hotspots", datadir=file.path(wd, "data/chronosphere"), verbose=FALSE)
+myers <- as(myers, "Spatial")
 
 ################################################################################
 # 2. Explicit WGS 84 CRS