+++ NEWS.md
@@ -1,6 +1,29 @@
 # Change log of the R package 'rgplates'
 
+# rgplates 0.3.0 - 2023-06-29 
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8098723.svg)](https://doi.org/10.5281/zenodo.8098723)
+
+### Added 
+- The `platemodel` class now accepts multiple feature collections. These can be recontructed by setting the `x` argument of `recontruct()`.  
+
+### Changed
+- The interface of the `reconstruct()` function now matches that of the GPlates web service. Use `"static_polygons"` instead of `"plates"`. 
+- The `plateperiod` argument of `reconstruct()` now defaults to `TRUE`
+- Documentation to reflect models avialble through the GPlates web service
+- Changed default method of online reconstruction to `"MERDITH2021"`
+
+### Removed
+- unnecessary imports from `utils`
+
+* * *
+
 # rgplates 0.2.1 - 2023-03-08 
+
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8094071.svg)](https://doi.org/10.5281/zenodo.8094071)
+
+
 ### Added
 - the `mapedge()` function
 - function website assets
@@ -11,6 +34,13 @@
 * * *
 
 # rgplates 0.2.0 - 2022-09-19
+
+
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8094042.svg)](https://doi.org/10.5281/zenodo.8094042)
+
+
+
 ### Added
 - Package now depends on the sustainable 'sf'-based interface to GPLates
 - Defense against non-integer reconstruction dates for the online method.
@@ -33,13 +63,18 @@
 * * *
 
 # rgplates 0.1.0 - 2021-05-11 
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8093991.svg)](https://doi.org/10.5281/zenodo.8093991)
+
 ### Changed
 - Bug fix for spaces in usernames on Windows machines.
 
 
 * * *
 
-# rgplates 0.1.0 - 2021-05-11 
+# rgplates 0.1.0 - 2021-05-09 
+
+
 ### Added 
 - material copied over from chronosphere 0.4.1 
 - reconstruct(model=NULL) argumentation.